pyproject.toml
[tool.poetry]
name = "volume-rendering-plot"
version = "0.1.0"
description = ""
authors = ["Rintarooo <310rnomeado@gmail.com>"]
readme = "README.md"
packages = [{include = "volume_rendering_plot"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "^1.26.0"
matplotlib = "^3.8.0"
plotly = "^5.17.0"
scipy = "^1.11.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
